requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
Unipath==1.0
dj-database-url==0.2.2
dj-static==0.0.5
django-decouple==2.1
django-filter==0.7
djangorestframework==2.3.12
gunicorn==18.0
psycopg2==2.5.2
pystache==0.5.3
pytz==2013.9
static==1.0.2
